Use our own version of texindex
Some versions of texindex don't support 4-arg \entry, which we need to
have subtopics show up nicely in the indices.

This is based on texindex from texinfo 6.5 and modified by me to
support 4-arg \entry macro.  This is based on work done with this for
maxima.

I'm using this because texindex 6.5 doesn't handle 4-arg \entry.  Some
tests show that the last version where this worked was 5.2, which was
the last version written in C.

For what ever reason, gitlab CI claims it's texindex 6.5, but
actually handles 4-arg entry.  We're using our own to make it
consistent everywhere.