Update links
We update many of the links with direct replacements if possible.
Sometimes it was just easier to point to the cliki page on the same
topic.

Finally, a few links are just broken, in which case we commented out
the link.  IF we can't find a new url for it, we'll remove these
eventually.