Use our version of texinfo.tex
When running makeinfo, set TEXINPUTS to point to include a path to our
version of texinfo.tex.  This is mostly for local building where the
version that might not be for texinfo 6.7.  Hence set path so we pick
up our version.

If the get the wrong version, we'll get errors about undefined
sequences related to @subentry.