Fix #20: Highlight navbar item
When we're on a particular page, highlight the navbar item to show
that we're on that page (or subpage).

To do this, when the navbar iframe is loaded from the main frame, a
query string is passed to the iframe indicating which item to
highlight.  The navbar iframe looks at the query string and finds the
entry in the list and highlights that entry.

Currently, highlighting just means make the item bold with a slightly
larger font.