Generate docs
The published site no longer has the docs.  So, doing an experiment
here to make the always available.  Add gen directory to cache.  The
docs install to the gen directory.  Then copy the gen directory to the
public directory.

Since the make files have changed, the docs should be generated.