Use webfont so page looks the same everywhere.
Well, only tested with Chrome and Firefox.

Somewhat arbitrarily choose Roboto and Roboto Mono because Roboto is a
san-serif font that we were already using and because it has a
monospaced version that matches (more or less).

Also set the line spacing to 1.5 so it's not too cramped for mobile
and desktop.

Let everything inherit the font-family from html instead of specifying
a font-family for each header type.  (We were using the same set
before anyway.)